NOJIRA update the setup and deploy tasks

+++ lib/tasks/setup.rake
@@ -5,7 +5,10 @@ namespace :setup do
     paths.each do |path|
       FileUtils.remove_dir(path)
     end
-    Rake::Task["db:drop"].invoke
+    # drop actually deletes the .sqlite3 file, which is not what we want on test
+    unless ENV['RAILS_ENV'].eql? "test"
+      Rake::Task["db:drop"].invoke
+    end
     Rake::Task["db:setup"].invoke
     Rake::Task["db:seed"].invoke
     Rake::Task["db:populate"].invoke
